Moved the test to uprade under observability
Moved the test under test_upgrade_observability
and fixed the dependecies.
