+++ tests/observability/storage/test_hpp_observability.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from tests.utils import validate_metrics_value
+from tests.observability.utils import validate_metrics_value
 from utilities.constants import (
     HOSTPATH_PROVISIONER_OPERATOR,
     TIMEOUT_2MIN,
